Refactor device join flow UI
