+++ src/components/device/index.ts
@@ -1,8 +1,8 @@
 export { default as PairedDevicesPanel } from './PairedDevicesPanel'
-export { default as DeviceStatus } from './DeviceStatus'
 export { default as Permissions } from './Permissions'
 export { default as Rules } from './Rules'
 export { default as Welcome } from './welcome'
 export { default as ThisDeviceCard } from './ThisDeviceCard'
 export { default as DeviceSettingsSheet } from './DeviceSettingsSheet'
+export { default as JoinExistingSpaceDialog } from './JoinExistingSpaceDialog'
 export { default as UnpairAlertDialog } from './UnpairAlertDialog'
